Add a "share this page" link.
This will help people share a permalink to a specific godecl query.
Typing in the input box clears the query. Clicking the permalink
navigates there instantly (without reloading the page unnecessary).

Give fixed amount of minimum space for output and permalink, so that
the footer doesn't move up and down when output size changes.