Show random initial example; add "random" link.
Since godecl is more of a spoof of cdecl rather than an actual useful
tool (since Go syntax is so easy to read and write), most people are
likely to just visit it for fun. They're unlikely to type stuff in the
input box.

Make the page more fun and interesting for them by showing random
initial examples. Do that on every page load, and also whenever they
click "random" link (without refreshing the page needlessly).